资源新增

+++ haozi-account/src/main/java/com/haozi/account/dao/po/RoleResources.java
@@ -1,14 +1,16 @@
 package com.haozi.account.dao.po;
 
 import com.baomidou.mybatisplus.annotation.TableName;
+
 import java.io.Serializable;
 import java.time.LocalDateTime;
+
 import lombok.Getter;
 import lombok.Setter;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author haozi
@@ -21,9 +23,9 @@ public class RoleResources implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-      private Integer roleId;
+    private Long roleId;
 
-      private Integer resourceId;
+    private Long resourceId;
 
     private LocalDateTime createTime;
